Refactor abstract image animator
Refactor `AbstractImageAnimator` to be memory-efficient. Previously,
using `ArtistAnimation` every image has to be stored in memory (a list),
as a result of which there will be a huge memory hog. Now, it just uses
`FuncAnimation` and only one image at a time is stored, and thus
animation is generated iteratively.